requirements.txt
Django==5.1
djangorestframework
requests
python-decouple
pre-commit
matplotlib
pytz
redis
django-redis
pandas